pyproject.toml
[project]
name = "mcp-qa"
version = "0.1.0"
description = "MCP Server para análisis de contratos Swagger/OpenAPI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.14.4",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]
